backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
gunicorn==20.1.0
python-dotenv==0.19.0
flake8==6.0.0
django-filter==23.2
isort==5.12.0
drf_extra_fields==3.6.1
django_colorfield==0.9.0
django-cors-headers==3.13.0
